+++ problem-spring-web-autoconfigure/src/main/java/org/zalando/problem/spring/web/autoconfigure/ExceptionHandling.java
@@ -8,6 +8,6 @@ import static org.apiguardian.api.API.Status.INTERNAL;
 
 @API(status = INTERNAL)
 @ControllerAdvice
-final class ExceptionHandling implements ProblemHandling {
+public final class ExceptionHandling implements ProblemHandling {
 
 }
