+++ problem-spring-web-autoconfigure/src/test/java/org/zalando/problem/spring/web/autoconfiguretests/CustomWebMvcApplicationTest.java
@@ -1,32 +1,27 @@
-package org.zalando.problem.spring.web.autoconfigure;
+package org.zalando.problem.spring.web.autoconfiguretests;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import org.zalando.problem.jackson.ProblemModule;
-import org.zalando.problem.spring.web.advice.AdviceTrait;
 import org.zalando.problem.violations.ConstraintViolationProblemModule;
 
 import javax.annotation.Nullable;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.NONE;
 
-@SpringBootTest(webEnvironment = NONE)
-@EnableAutoConfiguration
-final class NonWebApplicationTest {
+@SpringBootTest
+@EnableAutoConfiguration(exclude = SecurityAutoConfiguration.class)
+@EnableWebMvc
+final class CustomWebMvcApplicationTest {
 
     @Configuration
     static class TestApplication {
 
-    }
-
-    @Test
-    void shouldNotConfigureDefaultExceptionHandling(
-            @Autowired @Nullable final AdviceTrait trait) {
-        assertThat(trait).isNull();
     }
 
     @Test
