+++ problem-spring-web-autoconfigure/src/test/java/org/zalando/problem/spring/web/autoconfiguretests/WebApplicationTest.java
@@ -1,4 +1,4 @@
-package org.zalando.problem.spring.web.autoconfigure;
+package org.zalando.problem.spring.web.autoconfiguretests;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -6,18 +6,16 @@ import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import org.zalando.problem.jackson.ProblemModule;
+import org.zalando.problem.spring.web.advice.AdviceTrait;
+import org.zalando.problem.spring.web.autoconfigure.ExceptionHandling;
 import org.zalando.problem.violations.ConstraintViolationProblemModule;
-
-import javax.annotation.Nullable;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest
 @EnableAutoConfiguration(exclude = SecurityAutoConfiguration.class)
-@EnableWebMvc
-final class CustomWebMvcApplicationTest {
+final class WebApplicationTest {
 
     @Configuration
     static class TestApplication {
@@ -25,15 +23,21 @@ final class CustomWebMvcApplicationTest {
     }
 
     @Test
-    void shouldNotConfigureProblemModule(
-            @Autowired @Nullable final ProblemModule module) {
-        assertThat(module).isNull();
+    void shouldConfigureExceptionHandling(
+            @Autowired final AdviceTrait trait) {
+        assertThat(trait).isExactlyInstanceOf(ExceptionHandling.class);
     }
 
     @Test
-    void shouldNotConfigureConstraintViolationProblemModule(
-            @Autowired @Nullable final ConstraintViolationProblemModule module) {
-        assertThat(module).isNull();
+    void shouldConfigureProblemModule(
+            @Autowired final ProblemModule module) {
+        assertThat(module).isNotNull();
+    }
+
+    @Test
+    void shouldConfigureConstraintViolationProblemModule(
+            @Autowired final ConstraintViolationProblemModule module) {
+        assertThat(module).isNotNull();
     }
 
 }
