+++ problem-spring-web-autoconfigure/src/test/java/org/zalando/problem/spring/web/autoconfiguretests/security/NoConfigurerSecurityTest.java
@@ -1,13 +1,10 @@
-package org.zalando.problem.spring.web.autoconfigure.security;
+package org.zalando.problem.spring.web.autoconfiguretests.security;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.web.servlet.MockMvc;
 
@@ -19,12 +16,12 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @SpringBootTest
 @AutoConfigureMockMvc
 @DirtiesContext
-final class SecurityTest {
+final class NoConfigurerSecurityTest {
     @Autowired
     private MockMvc mockMvc;
 
     @Test
-    void shouldConfigureExceptionHandling() throws Exception{
+    void shouldConfigureExceptionHandling() throws Exception {
         mockMvc.perform(get("/not-exists-url"))
                 .andExpect(status().isUnauthorized())
                 .andExpect(jsonPath("$.status", is(401)))
@@ -34,12 +31,5 @@ final class SecurityTest {
 
     @SpringBootApplication
     static class TestApp {
-        @Configuration(proxyBeanMethods = false)
-        class WebSecurityConfig extends WebSecurityConfigurerAdapter {
-            @Override
-            protected void configure(HttpSecurity http) throws Exception {
-                http.authorizeRequests().anyRequest().authenticated();
-            }
-        }
     }
 }
